==> notebooks/0-creacion-vector-objetivo.ipynb <==
import pandas as pd
import numpy as np
from glob import glob
from tqdm.notebook import tqdm

BASE = "/media/giani/Gianicosas/Magister/Proyecto/raw_data"

# Proceso de matricula

## Datos

Se comparará la cantidad de alumnos por carrera por año para validación de datos

# Fuente: https://www.mifuturo.cl/bases-de-datos-de-matriculados/
proceso = pd.read_csv(f"{BASE}/OFICIAL_WEB_PROCESO_MAT_2007_AL_2022_24_06_2022_E.csv", encoding="latin-1", sep=";")
proceso.head()

print(list(proceso.columns)), proceso.shape

proceso["AÑO"].value_counts().sort_index(ascending=False)

# Matriculados

### Todas las columnas
```python
# pd.read_csv(f"{BASE}/20220719_Matrícula_Ed_Superior_2022_PUBL_MRUN.csv", nrows=5, sep=";").columns
[
    'cat_periodo', 'id', 'codigo_unico', 'mrun', 'gen_alu', 'fec_nac_alu', 'rango_edad', 'anio_ing_carr_ori',
    'sem_ing_carr_ori', 'anio_ing_carr_act', 'sem_ing_carr_act', 'anio_ing_carr_act', 'tipo_inst_1', 'tipo_inst_2', 'tipo_inst_3',
    'cod_inst', 'nomb_inst', 'cod_sede', 'nomb_sede', 'cod_carrera', 'nomb_carrera', 'modalidad', 'jornada',
    'version', 'tipo_plan_carr', 'dur_estudio_carr', 'dur_proceso_tit', 'dur_total_carr', 'region_sede',
    'provincia_sede', 'comuna_sede', 'nivel_global', 'nivel_carrera_1', 'nivel_carrera_2', 'requisito_ingreso',
    'vigencia_carrera', 'formato_valores', 'valor_matricula', 'valor_arancel', 'codigo_demre', 'area_conocimiento',
    'cine_f_97_area', 'cine_f_97_subarea', 'area_carrera_generica', 'cine_f_13_area', 'cine_f_13_subarea',
    'acreditada_carr', 'acreditada_inst', 'acre_inst_desde_hasta', 'acre_inst_anio', 'costo_proceso_titulacion',
    'costo_obtencion_titulo_diploma', 'forma_ingreso'
]
```

### Cols para determinar vector objetivo
```python
['cat_periodo', 'codigo_unico', 'mrun', 'anio_ing_carr_ori', 'anio_ing_carr_act', 'tipo_inst_1', 'dur_estudio_carr']
```

### Tipos de instituciones
```python
# Salida: ['Universidades', 'Institutos Profesionales', 'Centros de Formación Técnica']
pd.concat([
    pd.read_csv(f, sep=";", usecols=["tipo_inst_1"])
    for f in tqdm(glob(f"{BASE}/20220719_Matrícula_Ed_Superior_*.csv"), total=16)
])["tipo_inst_1"].unique()
instituciones = ['Institutos Profesionales', 'Centros de Formación Técnica']
```

## Lectura de datos matriculados

Fuente: https://datosabiertos.mineduc.cl/matricula-en-educacion-superior/

def iter_df_read(file, col_name, condition_list, chunksize=1000, **kwargs):
    iter_csv = pd.read_csv(file, iterator=True, chunksize=chunksize, **kwargs)
    return pd.concat([chunk[chunk[col_name].isin(condition_list)] for chunk in iter_csv])

cols_time = ['cat_periodo', 'codigo_unico', 'mrun', 'anio_ing_carr_ori', 'anio_ing_carr_act', 'tipo_inst_1', 'dur_estudio_carr']
instituciones = ['Institutos Profesionales', 'Centros de Formación Técnica']

matriculados = pd.concat([
    iter_df_read(f, "tipo_inst_1", instituciones, sep=";", usecols=cols_time)
    for f in tqdm(glob(f"{BASE}/post_2015/20220719_Matrícula_Ed_Superior_*.csv"), total=8)
])
matriculados.head()

### Información del DataFrame

matriculados.shape

matriculados.info()

matriculados.isna().sum()

### Pre procesamiento: Cursando carrera

Se crea columna `cursando`, la cual indica:
- `1` si los años transcurridos desde el año de ingreso a la carrera hasta el año 2022 es mayor a la duración de la carrera en años.
- `0` en caso contrario

Esto se utilizará más adelante en la construcción del vector objetivo; Se hace una selección de datos donde solo se selecciona a aquellos que tengan `cursando==0`, para medir inicialmente deserción solamente en aquellos que no se encuentren ya en la ventana de duración de la carrera.

matriculados = matriculados.dropna(subset=["mrun"]).reset_index(drop=True)
matriculados["cursando"] = np.where(
    pd.Series([2022] * matriculados.shape[0]) - matriculados["anio_ing_carr_ori"] > matriculados["dur_estudio_carr"] / 2, 0, 1
)

matriculados["cursando"].value_counts("%")

### Validación: Cantidad de matriculados por carrera por año

Se compara con la base de datos agregados

anios_proceso = sorted(proceso["AÑO"].unique(), reverse=True)[:8]
assert sorted(matriculados["cat_periodo"].unique(), reverse=True) == [int(a.split("_")[1]) for a in anios_proceso]

# Cantidad de matriculados por año y carrera
mat_por_carrera = matriculados.groupby(
    ["cat_periodo", "codigo_unico"]
).agg({"mrun": pd.Series.nunique}).rename(columns={"mrun": "cant_matriculados_carrera_anio"}).reset_index().sort_values(by=["cant_matriculados_carrera_anio"])
mat_por_carrera["cant_matriculados_carrera_anio"].describe()

proceso[
    proceso["AÑO"].isin(anios_proceso)
].groupby([
    "AÑO", "CÓDIGO CARRERA"
]).agg({
    "TOTAL MATRICULADOS": "sum"
}).reset_index()["TOTAL MATRICULADOS"].describe()

- La distribución de matriculados por carrera por año difiere entre los datos agregados y los desagregados

mat_por_carrera["codigo_unico"].nunique(), proceso[(proceso["AÑO"].isin("MAT_" + mat_por_carrera["cat_periodo"].astype(str)))]["CÓDIGO CARRERA"].nunique()

- Los datos agregados tienen 16.000 carreras más que los desagregados.
    - Ejemplo de carrera no encontrada en datos desagregados:

print(list(set(proceso[(proceso["AÑO"].isin("MAT_" + mat_por_carrera["cat_periodo"].astype(str)))]["CÓDIGO CARRERA"].unique())-set(mat_por_carrera["codigo_unico"].unique()))[:10])

proceso[
    (proceso["AÑO"].isin("MAT_" + mat_por_carrera["cat_periodo"].astype(str)))
    & (proceso["CÓDIGO CARRERA"]=='I92S1C173J3V1')
][["AÑO", "TOTAL MATRICULADOS", "CÓDIGO CARRERA"]]

mat_por_carrera[mat_por_carrera["codigo_unico"] == "I92S1C173J3V1"]

- Dentro de las carreras que sí coinciden en ambos sets para el mismo periodo, se explora la diferencia de matriculados

coinciden = proceso[
    (proceso["AÑO"].isin("MAT_" + mat_por_carrera["cat_periodo"].astype(str)))
    & (proceso["CÓDIGO CARRERA"].isin(mat_por_carrera["codigo_unico"]))
].reset_index(drop=True).sort_values("TOTAL MATRICULADOS")

cant_mat_dif = coinciden[
    coinciden["TOTAL MATRICULADOS"] != mat_por_carrera["cant_matriculados_carrera_anio"].reset_index(drop=True).values
][["AÑO", "CÓDIGO CARRERA", "TOTAL MATRICULADOS"]].rename(columns={"CÓDIGO CARRERA": "codigo_unico"})
cant_mat_dif["cat_periodo"] = cant_mat_dif["AÑO"].str.split("_", expand=True)[1].astype(int)

difieren = pd.merge(cant_mat_dif, mat_por_carrera)
difieren["diferencia"] = difieren["TOTAL MATRICULADOS"] - difieren["cant_matriculados_carrera_anio"]
difieren["diferencia"].describe()

- En general la cantidad de matriculados coincide, habiendo una diferencia máxima de 6 alumnos. Por tanto, se concluye que la gran diferencia en las distribuciones de matriculados por carrera por año se debe a que en el set de datos desagregados falta incluir muchas carreras.

### Validación: Código único

- Se corrobora que un código único puede estar asociado a distintos nombres de carrera, pero se asumirá que todos los nombres diferentes hacen referencia a la misma.
- Todos los códigos únicos repetidos pertenecen a la misma institución
- Se observa un ejemplo

proceso.groupby(["CÓDIGO CARRERA"]).agg({"NOMBRE INSTITUCIÓN": pd.Series.nunique, "NOMBRE CARRERA": pd.Series.nunique}).reset_index().sort_values(by=["NOMBRE CARRERA"], ascending=False).head()

proceso[proceso["CÓDIGO CARRERA"]=="I71S1C221J2V1"][["NOMBRE INSTITUCIÓN", "NOMBRE CARRERA", "CÓDIGO CARRERA"]].drop_duplicates()

## Operacionalización de datos

### Alumno desertor
- Tiene que haber pasado un tiempo mayor a la duración de la carrera
- El último semestre cursado tiene que ser menor a la cantidad de semestres de la carrera (no se cuenta con este dato)
- Considerar todo para el mismo codigo de carrera cada vez

## Vector Objetivo: Primera iteración
- Se considera para esta primera iteración aquellos alumnos que ya hayan superado la ventana de tiempo de duración de la carrera.
- Como no se tiene el dato de qué número de semestre está cursando el alumno (considerando el total de semestres disponibles), se asumirá al menos en esta primera etapa que los alumnos no han repetido semestres/años. Esto claramente subestimará la cantidad de desertores, ya que un alumno puede haber completado la cantidad de años teóricos cursados, sin haber completado la carrera (por repetición de un ramo, congelamiento, etc), lo que lo calificaría como desertor.

# Algunos aparecen 2 veces probablemente porque se matriculan 2 veces al año, por eso se elimina duplicados
termino_duracion_carrera = matriculados[matriculados["cursando"]==0][["cat_periodo", "codigo_unico", "mrun", "anio_ing_carr_ori"]].drop_duplicates()

# Se corrobora que solo hay 1 año de ingreso a una carrera por alumno por carrera
ai = termino_duracion_carrera.groupby(["codigo_unico","mrun"]).agg({"anio_ing_carr_ori": pd.Series.nunique}).reset_index()
ai["anio_ing_carr_ori"].describe()

ai.shape, ai[ai["anio_ing_carr_ori"]>1].shape, 31971/1489620 * 100

ai.sort_values(by="anio_ing_carr_ori", ascending=False)

matriculados[(matriculados["codigo_unico"]=="I113S10C430J2V1") & (matriculados["mrun"]==24146362.0)]

matriculados[(matriculados["codigo_unico"].isin(["I113S10C430J2V1", "I214S20C19J1V2"])) & (matriculados["mrun"].isin([24146362.0, 15903180.0]))].sort_values(by=["codigo_unico"])

matriculados[matriculados["mrun"].isin([24146362.0])]

termino_anio_ori_unico = termino_duracion_carrera[
    (termino_duracion_carrera["codigo_unico"].isin(ai[ai["anio_ing_carr_ori"]>1]["codigo_unico"]))
    & (termino_duracion_carrera["mrun"].isin(ai[ai["anio_ing_carr_ori"]>1]["mrun"]))
]

t = termino_duracion_carrera.groupby(["codigo_unico", "mrun"]).size().reset_index(name="cantidad_de_anios_cursados_por_carrera")
t[(t["codigo_unico"]=="I113S10C430J2V1") & (t["mrun"]==24146362.0)]

t2 = termino_anio_ori_unico.groupby(["codigo_unico", "mrun"]).size().reset_index(name="cantidad_de_anios_cursados_por_carrera")
t2[(t2["codigo_unico"]=="I113S10C430J2V1") & (t["mrun"]==24146362.0)]

anios_matriculado = pd.merge(
    pd.merge(
        pd.merge(
            termino_duracion_carrera.groupby( 
                ["codigo_unico", "mrun"]
            ).size().reset_index(name="cantidad_de_anios_cursados_por_carrera"),
            matriculados[["codigo_unico", "dur_estudio_carr"]].drop_duplicates(),
            how="left"
        ),
        termino_duracion_carrera.groupby(["codigo_unico","mrun"])["cat_periodo"].max().reset_index().rename(columns={"cat_periodo": "ultimo_anio_matricula"}),
        how="left"
    ),
    termino_duracion_carrera.groupby(["codigo_unico","mrun"])["anio_ing_carr_ori"].max().reset_index().rename(columns={"anio_ing_carr_ori": "primer_anio_matricula"}),
    how="left"
)

anios_matriculado["duracion_carrera_anios"] = anios_matriculado["dur_estudio_carr"] / 2

anios_matriculado.drop(columns=["dur_estudio_carr"]).sort_values(by=["cantidad_de_anios_cursados_por_carrera"], ascending=False)

# Se debe marcar coo desertor al alumno en el último año cursado

#### DEBUGS

1589215 

matriculados["mrun"].value_counts().head()

matriculados[matriculados["mrun"]==7085589.0].sort_values(by=["mrun", "cat_periodo"]).head(10)
