==> notebooks/1-creacion-vector-objetivo.ipynb <==
import pandas as pd
import numpy as np
from glob import glob
from tqdm.notebook import tqdm
from utils import iter_df_read

# Modificar por ruta local de almacenamiento de datos
BASE = "/media/giani/Gianicosas/Magister/Proyecto/"

# Matriculados

### Todas las columnas
```python
# pd.read_csv(f"{BASE}/20220719_Matrícula_Ed_Superior_2022_PUBL_MRUN.csv", nrows=5, sep=";").columns
[
    'cat_periodo', 'id', 'codigo_unico', 'mrun', 'gen_alu', 'fec_nac_alu', 'rango_edad', 'anio_ing_carr_ori',
    'sem_ing_carr_ori', 'anio_ing_carr_act', 'sem_ing_carr_act', 'anio_ing_carr_act', 'tipo_inst_1', 'tipo_inst_2', 'tipo_inst_3',
    'cod_inst', 'nomb_inst', 'cod_sede', 'nomb_sede', 'cod_carrera', 'nomb_carrera', 'modalidad', 'jornada',
    'version', 'tipo_plan_carr', 'dur_estudio_carr', 'dur_proceso_tit', 'dur_total_carr', 'region_sede',
    'provincia_sede', 'comuna_sede', 'nivel_global', 'nivel_carrera_1', 'nivel_carrera_2', 'requisito_ingreso',
    'vigencia_carrera', 'formato_valores', 'valor_matricula', 'valor_arancel', 'codigo_demre', 'area_conocimiento',
    'cine_f_97_area', 'cine_f_97_subarea', 'area_carrera_generica', 'cine_f_13_area', 'cine_f_13_subarea',
    'acreditada_carr', 'acreditada_inst', 'acre_inst_desde_hasta', 'acre_inst_anio', 'costo_proceso_titulacion',
    'costo_obtencion_titulo_diploma', 'forma_ingreso'
]
```

### Cols para determinar vector objetivo
```python
['cat_periodo', 'codigo_unico', 'mrun', 'anio_ing_carr_ori', 'anio_ing_carr_act', 'tipo_inst_1', 'dur_estudio_carr']
```

### Tipos de instituciones
```python
# Salida: ['Universidades', 'Institutos Profesionales', 'Centros de Formación Técnica']
pd.concat([
    pd.read_csv(f, sep=";", usecols=["tipo_inst_1"])
    for f in tqdm(glob(f"{BASE}/20220719_Matrícula_Ed_Superior_*.csv"), total=16)
])["tipo_inst_1"].unique()
instituciones = ['Institutos Profesionales', 'Centros de Formación Técnica']
```

## Lectura de datos matriculados

Fuente: https://datosabiertos.mineduc.cl/matricula-en-educacion-superior/

- Se considera solo matriculados en IP y CFT
- Se elimina los registros con `mrun` nulo
- Se considera solo registros desde el 2015
- Se reemplaza los datos `9995` por `np.nan`

cols_time = [
    'cat_periodo', 'codigo_unico', 'mrun', 'anio_ing_carr_ori', 'anio_ing_carr_act',
    'tipo_inst_1', 'dur_estudio_carr', 'cod_carrera', 'cod_inst', 'cod_sede', 'version'
]
instituciones = ['Institutos Profesionales', 'Centros de Formación Técnica']

matriculados = pd.concat([
    iter_df_read(f, "tipo_inst_1", instituciones, sep=";", usecols=cols_time)
    for f in tqdm(
        glob(f"{BASE}/raw_data/post_2015/20220719_Matrícula_Ed_Superior_*.csv"),
        total=8
    )
]).dropna(subset="mrun")

matriculados.loc[:, ['anio_ing_carr_ori', 'anio_ing_carr_act']] = matriculados[
    ['anio_ing_carr_ori', 'anio_ing_carr_act']
].replace(
    [9995], [np.nan] 
)

print(f"Cantidad de registros: {matriculados.shape[0]}\n")
print(f"Cantidad de nulos: \n{matriculados.isna().sum()}")
matriculados.head()

## Pre procesamiento: Año ingreso carrera actual

En los años anteriores a 2018 este dato está nulo en muchos registros. Por ahora se reemplaza el valor por el del año de ingreso a carrera original.

matriculados['anio_ing_carr_act'] = matriculados['anio_ing_carr_act'].fillna(matriculados['anio_ing_carr_ori'])
matriculados['anio_ing_carr_act'].isna().sum()

matriculados['anio_ing_carr_act'].isin([9998, 9999]).value_counts()

## Pre procesamiento; Año ingreso carrera original
Se informa desde la contraparte que se debe utilizar esta columna como referencia de año de ingreso para tomar la cantidad de años cursados

matriculados['anio_ing_carr_ori'].isin([9998, 9999]).value_counts("%")

# Por ahora se decide eliminar estos registros
print("Cantidad de matrículas considerando todos los registros: ", matriculados.shape[0])
matriculados = matriculados[~matriculados['anio_ing_carr_ori'].isin([9998, 9999])]
print("Cantidad de matriculas omitiendo año de ingreso 9998 o 9999: ", matriculados.shape[0])

### Mínimo año de ingreso carrera original (primera vez que ingresó a educación superior)

min_anio_ing_ori = matriculados.groupby(["mrun"]).agg({"anio_ing_carr_ori": "min"}).reset_index().rename(columns={"anio_ing_carr_ori": "min_anio_ing_carr_ori"})
min_anio_ing_ori.head()

if "min_anio_ing_carr_ori" not in matriculados.columns:
    shape_or = matriculados.shape[0]
    matriculados = pd.merge(
        matriculados,
        min_anio_ing_ori,
        how="left"
    )
    
    assert shape_or == matriculados.shape[0]

## Pre procesamiento: Cursando carrera

Se crea columna `cursando`, la cual indica:
- `1` si los años transcurridos desde el año de ingreso a la carrera hasta el año 2022 es menor o igual a la duración de la carrera en años.
- `0` en caso contrario (debió ya terminar la carrera)

Esto se utilizará más adelante en la construcción del vector objetivo; Se hace una selección de datos donde solo se selecciona a aquellos que tengan `cursando==0`, para medir inicialmente deserción solamente en aquellos que no se encuentren ya en la ventana de duración de la carrera.

matriculados = matriculados.dropna(subset=["mrun"]).reset_index(drop=True)
matriculados["cursando"] = np.where(
    pd.Series([2022] * matriculados.shape[0]) - matriculados["anio_ing_carr_ori"]
    <= matriculados["dur_estudio_carr"] / 2, 1, 0
)

matriculados["cursando"].value_counts("%")

# Se conserva las combinaciones únicas carrera-alumno de quienes ingresaron a la carrera "hace más tiempo" de lo que dura la carrera
termino_duracion_carrera = matriculados[matriculados["cursando"]==0][
    ["codigo_unico", "mrun"]
].drop_duplicates().reset_index(drop=True)

print(f"Combinaciones únicas alumno-carrera que pasó ya el tiempo de duración de la carrera: ", termino_duracion_carrera[["codigo_unico", "mrun"]].drop_duplicates().shape[0])

# Titulados

Fuente: https://datosabiertos.mineduc.cl/titulados-en-educacion-superior/

### Todas las columnas
```python
[
    'cat_periodo', 'codigo_unico', 'mrun', 'gen_alu', 'fec_nac_alu', 'rango_edad', 'anio_ing_carr_ori',
    'sem_ing_carr_ori', 'anio_ing_carr_act', 'sem_ing_carr_act', 'nomb_titulo_obtenido',
    'nomb_grado_obtenido', 'fecha_obtencion_titulo', 'tipo_inst_1', 'tipo_inst_2', 'tipo_inst_3',
    'cod_inst', 'nomb_inst', 'cod_sede', 'nomb_sede', 'cod_carrera', 'nomb_carrera', 'nivel_global',
    'nivel_carrera_1', 'nivel_carrera_2', 'dur_estudio_carr', 'dur_proceso_tit', 'dur_total_carr',
    'region_sede', 'provincia_sede', 'comuna_sede', 'jornada', 'modalidad', 'version', 'tipo_plan_carr',
    'area_cineunesco', 'area_cine_f_97', 'subarea_cine_f_97', 'area_cine_f_13', 'subarea_cine_f_13',
    'area_carrera_generica_n'
]
```

### Cols para determinar vector objetivo
```python
['cat_periodo', 'codigo_unico', 'mrun', 'tipo_inst_1', 'fecha_obtencion_titulo']
```

cols_tit = [
    'cat_periodo', 'codigo_unico', 'mrun', 'tipo_inst_1', 'fecha_obtencion_titulo',
     'cod_carrera', 'cod_inst', 'cod_sede', 'version'
]

titulados = pd.concat([
    iter_df_read(f, "tipo_inst_1", instituciones, sep=";", usecols=cols_tit)
    for f in tqdm(
        glob(f"{BASE}/raw_data/titulados_post_2015/20220804_Titulados_Ed_Superior_*.csv"),
        total=7
    )
]).dropna(subset="mrun")

print(titulados.shape)
titulados.head()

titulados.isna().sum()

# Unión de bases

merge = pd.merge(
    # Considero a todos finalmente porque puede haber desertado alguien sin que haya pasado la duración de su carrera.
    # Ej: Se matriculó el 2020, no se matriculó el 2021, y su carrera dura 3 años -> Este dato lo debo rescatar también
    # Se establece año de ingreso de carrera original desde 2015 para hacer el corte desde ese año
    # Se excluye a los que ingresaron el 2022
    matriculados[
        (matriculados["min_anio_ing_carr_ori"] >= 2015)
        & (matriculados["min_anio_ing_carr_ori"] < 2022)
    ][["codigo_unico", "mrun"]], 
    titulados[["codigo_unico", "mrun", "fecha_obtencion_titulo"]],
    how="left"
).sort_values(
    by=["fecha_obtencion_titulo"], ascending=False
).drop_duplicates(
    subset=["codigo_unico", "mrun"], keep="first" # Algunos aparecen con 2 fechas de titulación (approx 100), se deja la más reciente
)

termino_duracion_carrera.shape, titulados.shape, merge.shape

# Prueba haciendo cruce por combinación de variables de carrera, sin considerar versión

merge2 = pd.merge(
    matriculados[
        (matriculados["min_anio_ing_carr_ori"] >= 2015)
        & (matriculados["min_anio_ing_carr_ori"] < 2022)
    ][["codigo_unico", 'cod_carrera', 'cod_inst', 'cod_sede', "mrun"]], # Se conserva el código único de la base de matriculados para poder hacer los cruces correctos
    titulados[['cod_carrera', 'cod_inst', 'cod_sede', "mrun", "fecha_obtencion_titulo"]],
    how="left"
).sort_values(
    by=["fecha_obtencion_titulo"], ascending=False
).drop_duplicates(
    subset=['cod_carrera', 'cod_inst', 'cod_sede', "mrun"], keep="first"
)

termino_duracion_carrera.shape, titulados.shape, merge.shape

# Potenciales desertores
- No se han titulado, y ya ha pasado el tiempo de duración de su carrera
- Se debe corroborar luego si es que siguen estudiando; Puede que no se hayan titulado aún porque han alargado su tiempo de estudio respecto de la duración esperada

# Desertores (no se han titulado y ya ha pasado el tiempo para que se titulacen)
round(merge["fecha_obtencion_titulo"].isna().value_counts("%").rename(
    index={True: "Potencial desertor", False: "Titulado"}
) * 100, 1)

round(merge2["fecha_obtencion_titulo"].isna().value_counts("%").rename(
    index={True: "Potencial desertor", False: "Titulado"}
) * 100, 1)

# Serialización
- Se almacena la información sobre la fecha de titulación para carreras de CFT / IP, para hacer los cruces pertinentes con otros datos en otros notebooks

merge2.to_csv(f"{BASE}/tmp_data/fechas_titulacion.csv", index=False)
