==> notebooks/3-seleccion-atributos.ipynb <==
import pandas as pd
import numpy as np
from glob import glob
from tqdm.notebook import tqdm
from utils import iter_df_read, get_array_diff
import matplotlib.pyplot as plt
import seaborn as sns
plt.style.use("seaborn")
sns.set(font_scale=1.5)

# Modificar por ruta local de almacenamiento de datos
BASE = "/media/giani/Gianicosas/Magister/Proyecto/"

# Atributos

## Primera iteración
- **Notas de colegio**: Egresados Colegio, `PROM_NOTAS_ALU`.
- **Región del colegio**: Egresados Colegio, `COD_REG_RBD`.
- **Modalidad del colegio (científico humanista / técnico)**: Egresados Colegio, `COD_ENSE`.
- **Tipo de establecimiento (particular, subvencionado, municipal)**: Establecimientos, `cod_depe`.
- **Género**: Matriculados, `gen_alu`.
- **Edad**: Matriculados, `fec_nac_alu`.
- **Área del conocimiento de la carrera**: Matriculados, `area_conocimiento`.
- **Región de la institución de educación superior**: Matriculados, `region_sede`.

## Segunda iteración
- **Valor Arancel**: Matriculados, `valor_arancel` (Matriz 2).
- **Estado acreditación carrera**: Matriculados, `acreditada_carr` (Matriz 2).
- **Estado acreditación institución**: Matriculados, `acreditada_inst` (Matriz 2).
- **Cantidad de becas**: Becas y créditos, `BENEFICIO_BECA_FSCU` (Matriz 2).

## Tercera iteración
- **FSCU**: Tiene fondo solidario, `BENEFICIO_BECA_FSCU` (Matriz 3).
- **gratuidad**: Tiene gratuidad, `BENEFICIO_BECA_FSCU` (Matriz 3).
- **beca**: Tiene beca, `BENEFICIO_BECA_FSCU` (Matriz 3).

## Cuarta iteración
- **Jornada**: Diurno o vespertino (Matriz 4).
- **Duración carrera**: Duración en semestres de la carrera (Matriz 4).
- **Duración titulación**: Duración en semestres del proceso de titulación (Matriz 4).

Se corrige además el valor del arancel, para lo cual se tiene que consultar el formato del valor. **Nota**: En esta iteración también se agrega subtipo de institución, lo cual se encuentra en el notebook 6 de pre procesamiento.

## Quinta iteración
- **Decil**: Nivel socioeconomico (Matriz 5).

# Lectura datos

## Desertores

desertores_1 = pd.read_csv(f"{BASE}/tmp_data/desertores_1.csv.gz")
desertores_1.head()

## Matrículas

cols_mat = [
    'cat_periodo', 'codigo_unico', 'mrun', 'tipo_inst_1', 
    'gen_alu', 'fec_nac_alu', 'area_conocimiento', 'region_sede',
    "anio_ing_carr_ori",  "valor_arancel", "acreditada_carr", "acreditada_inst",
    "jornada", "dur_estudio_carr", "dur_proceso_tit", "formato_valores"

]

instituciones = ['Institutos Profesionales', 'Centros de Formación Técnica']

matriculados = pd.concat([
    iter_df_read(f, "tipo_inst_1", instituciones, sep=";", usecols=cols_mat)
    for f in tqdm(
        glob(f"{BASE}/raw_data/post_2015/20220719_Matrícula_Ed_Superior_*.csv"),
        total=8
    )
]).dropna(subset="mrun").drop_duplicates().reset_index(drop=True)

matriculados = matriculados[
    (matriculados["anio_ing_carr_ori"] >= 2015)
    & (matriculados["anio_ing_carr_ori"] < 2022)
]

print(f"Cantidad de registros: {matriculados.shape[0]}\n")
print(f"Cantidad de nulos: \n{matriculados.isna().sum()}")
matriculados.head()

matriculados["formato_valores"] = matriculados["formato_valores"].fillna('Monto en Pesos')

mat_des = pd.merge(
    matriculados,
    desertores_1,
    how="left"
)

matriculados.shape, desertores_1.shape, mat_des.shape

mat_des[mat_des["desertor_1"].isna()]["cat_periodo"].value_counts()

mat_des = mat_des.dropna().reset_index(drop=True)

## Colegios

### Egresados EM

cols_egreso_col = [
    'MRUN', 'PROM_NOTAS_ALU', 'COD_REG_RBD', 'RBD', 'COD_ENSE', 'COD_GRADO', "AGNO"
]

egreso_colegio_todos = pd.concat([
    iter_df_read(f, "MRUN", mat_des["mrun"].unique(), sep=";", usecols=cols_egreso_col)
    for f in tqdm(
        glob(f"{BASE}raw_data/egresados_colegios/20220325_Notas_y_Egresados_Enseñanza_Media_*.csv"),
        total=20
    )
]).dropna(subset="MRUN").drop_duplicates().reset_index(drop=True).rename(
   columns={"MRUN": "mrun"}
)

egreso_colegio_todos["PROM_NOTAS_ALU"] = egreso_colegio_todos["PROM_NOTAS_ALU"].str.replace(",", ".").astype(float)

prom_notas = egreso_colegio_todos.groupby(["mrun"]).agg({
    "PROM_NOTAS_ALU": "mean"
}).reset_index().rename(columns={
    "PROM_NOTAS_ALU": "prom_notas_media"
})

egreso_colegio = pd.merge(
    egreso_colegio_todos,
    prom_notas,
    how="left"
).sort_values(
    by=["COD_GRADO"], ascending=False
).drop_duplicates(
    subset=["mrun"], keep="first"
).reset_index(drop=True)

print(egreso_colegio_todos.shape, egreso_colegio.shape)
print(f"Cantidad de registros: {egreso_colegio.shape[0]}\n")
print(f"Cantidad de nulos: \n{egreso_colegio.isna().sum()}")
egreso_colegio.head()

### Establecimientos
https://datosabiertos.mineduc.cl/directorio-de-establecimientos-educacionales/

- Se modificó los nombres de columnas manualmente en los casos que no coincidían.
- Además, en los archivos antiguos, se agregó la columna para el año.
- Los archivos que venían como xlsx se guardaron de nuevo como xls para facilitar la carga.

establecimientos_04_16 = pd.concat([
    pd.read_csv(f, sep=";").rename(columns={
        "RBD":"rbd",
        "COD_DEPE":"cod_depe",
        "AGNO": "agno",
        "ïagno": "agno"
    })
    for f in tqdm(
        glob(f"{BASE}raw_data/directorio_colegios/Directorio_oficial_*.csv"),
        total=13
    )
])[[
    "rbd", "cod_depe", "agno"
]].dropna()

establecimientos_04_16 = establecimientos_04_16[
    establecimientos_04_16["rbd"].isin(egreso_colegio["RBD"].unique())
].drop_duplicates().reset_index(drop=True)

print(f"Cantidad de registros: {establecimientos_04_16.shape[0]}\n")
print(f"Cantidad de nulos: \n{establecimientos_04_16.isna().sum()}")
establecimientos_04_16.head()

establecimientos_17_21 = pd.concat([
    pd.read_csv(f).rename(columns={
        "RBD":"rbd",
        "COD_DEPE":"cod_depe",
        "AGNO": "agno",
        "ïagno": "agno"
    })
    for f in tqdm(
        glob(f"{BASE}raw_data/directorio_colegios/Directorio_Oficial_*.csv"),
        total=5
    )
])[[
    "rbd", "cod_depe", "agno"
]].dropna()

establecimientos_17_21 = establecimientos_17_21[
    establecimientos_17_21["rbd"].isin(egreso_colegio["RBD"].unique())
].drop_duplicates().reset_index(drop=True)

print(f"Cantidad de registros: {establecimientos_17_21.shape[0]}\n")
print(f"Cantidad de nulos: \n{establecimientos_17_21.isna().sum()}")
establecimientos_17_21.head()

establecimientos_antiguos = pd.concat([
    pd.read_excel(f).rename(columns={
        "RBD":"rbd",
        "COD_DEPE":"cod_depe",
        "ANO": "agno"
    })
    for f in tqdm(
        glob(f"{BASE}raw_data/directorio_colegios/*.xls"),
        total=12
    )
])[[
    "rbd", "cod_depe", "agno"
]].dropna()

establecimientos_antiguos = establecimientos_antiguos[
    establecimientos_antiguos["rbd"].isin(egreso_colegio["RBD"].unique())
].drop_duplicates().reset_index(drop=True)

print(f"Cantidad de registros: {establecimientos_antiguos.shape[0]}\n")
print(f"Cantidad de nulos: \n{establecimientos_antiguos.isna().sum()}")
establecimientos_antiguos.head()

establecimientos_antiguos.isna().sum()

establecimientos = pd.concat([establecimientos_04_16, establecimientos_17_21, establecimientos_antiguos])
establecimientos.head()

sorted(establecimientos["agno"].unique())

colegios = pd.merge(
    egreso_colegio.rename(columns={
        "RBD":"rbd",
        "AGNO": "agno"
    }),
    establecimientos,
    how="left"
).drop_duplicates().reset_index(drop=True)

colegios.shape, egreso_colegio.shape

colegios.head()

colegios[colegios["cod_depe"].isna()].shape

### Becas
https://datosabiertos.mineduc.cl/asignaciones-de-becas-y-creditos-en-educacion-superior/

becas = pd.concat([
    pd.read_csv(f, sep=";")
    for f in tqdm(
        glob(f"{BASE}raw_data/becas_post_2015/Asignacion*.csv"),
        total=6
    )
]).drop(columns=[
    "TIPO_ALUMNO"
]).rename(columns={
    "ANIO_BENEFICIO": "cat_periodo",
    "MRUN": "mrun"
})

becas = becas[
    becas["mrun"].isin(matriculados["mrun"].unique())
].drop_duplicates().reset_index(drop=True)

print(f"Cantidad de registros: {becas.shape[0]}\n")
print(f"Cantidad de nulos: \n{becas.isna().sum()}")
becas.head()

# Segunda iteración
cant_becas = becas.groupby(["cat_periodo", "mrun"])["BENEFICIO_BECA_FSCU"].size().reset_index(name="cant_becas")
cant_becas.sort_values(by=["cant_becas"], ascending=False)

# Tercera iteración
becas["fscu"] = np.where(becas["BENEFICIO_BECA_FSCU"] == "FSCU", 1, 0)
becas["gratuidad"] = np.where(becas["BENEFICIO_BECA_FSCU"] == "GRATUIDAD", 1, 0)
becas["beca"] = np.where(~becas["BENEFICIO_BECA_FSCU"].isin(["FSCU", "GRATUIDAD"]), 1, 0)

# 1712062
print(becas["fscu"].value_counts())
print(becas["gratuidad"].value_counts())
print(becas["beca"].value_counts())

becas_merge = becas.drop(columns=["BENEFICIO_BECA_FSCU"]).drop_duplicates().groupby(["cat_periodo", "mrun"]).agg({
    "fscu": "sum",
    "gratuidad": "sum",
    "beca": "sum"
}).reset_index()
print(becas_merge.shape, becas.shape)

becas_merge["fscu"].max(), becas_merge["gratuidad"].max(), becas_merge["beca"].max(),

## Deciles

deciles = pd.read_csv(f"{BASE}raw_data/base_deciles.csv", sep=";").drop(columns=[
    "paste", "mrun.y", "año"
]).rename(columns={
    "mrun.x": "mrun"
})
print(deciles.shape)
print(deciles.isna().sum())
deciles.head()

# Unión de datos

## Unión información de matrícula y colegios

merge_1 = pd.merge(
    mat_des,
    colegios[["mrun", "COD_REG_RBD", "COD_ENSE", "prom_notas_media", "cod_depe"]].rename(columns={
        "COD_REG_RBD":"region_colegio",
        "COD_ENSE": "tipo_ensenanza_colegio", #CH, tecnico, etc
        "cod_depe": "dependencia_colegio" #municipal, subvencionado, etc
    }),
    how="left"
).drop_duplicates().reset_index(drop=True)

colegios.shape, mat_des.shape, merge_1.shape

merge_1.isna().sum()

## Anterior más información de becas

merge_2 = pd.merge(
    merge_1,
    becas_merge,
    how="left"
).drop_duplicates().reset_index(drop=True)

for c in ["fscu", "gratuidad", "beca"]:
    merge_2[c] = merge_2[c].fillna(0)

print(becas_merge.shape, merge_1.shape, merge_2.shape)

merge_2.isna().sum()

## Anterior más información de deciles

merge_3 = pd.merge(
    merge_2,
    deciles,
    how="left"
).drop_duplicates().reset_index(drop=True)

print(deciles.shape, merge_2.shape, merge_3.shape)

merge_3.isna().sum()

# Serialización

merge_3.to_csv(f"{BASE}work_data/matriz_desercion_5.csv.gz", index=False, compression="gzip")
