==> notebooks/6-preprocesamiento.ipynb <==
import pandas as pd
import numpy as np
import seaborn as sns
from datetime import date
from unidecode import unidecode
from tqdm.notebook import tqdm
from glob import glob

BASE = "/media/giani/Gianicosas/Magister/Proyecto/"

# Nuevos atributos

A los anteriores, se agrega sub tipo de institución: 
- CFT: Estatal / Privado
- IP: Presencia nacional / Privado

# Raw data

df = pd.read_csv(f"{BASE}work_data/matriz_desercion_5.csv.gz")
df.head()

## Selección periodo
- `cat_periodo` entre 2015 y 2019 inclusive. 

df = df[df["cat_periodo"] < 2020]

# Info solicitada para deciles
df_mrun = df[["cat_periodo", "mrun"]].drop_duplicates()
print(df_mrun.shape)

df_mrun.to_csv(f"{BASE}work_data/mrun_anio.csv", index=False)

# Nulos

df["fec_nac_alu"] = df["fec_nac_alu"].replace([190001], [np.nan])
df["anio_ing_carr_ori"] = df["anio_ing_carr_ori"].replace([9995], [np.nan])
df["acreditada_carr"] = df["acreditada_carr"].str.strip().replace(["", "SIN INFORMACION"], [np.nan, np.nan])
df["acreditada_inst"] = df["acreditada_inst"].str.strip().replace([""], [np.nan])
print(df.shape)
df = df.dropna(subset=[
    "fec_nac_alu", "anio_ing_carr_ori", "acreditada_carr", "acreditada_inst",
    "region_colegio", "tipo_ensenanza_colegio", "prom_notas_media", "dependencia_colegio"
])
print(df.shape)

# Pre procesamiento variables dependientes datos desagregados

## Diferencia año ingreso y último año de matrícula

g = df.groupby(["mrun", "codigo_unico"]).agg({
    "cat_periodo": "max",
    "anio_ing_carr_ori": "min"
}).reset_index()
g["dif_ing_ult_mat"] = g["cat_periodo"] - g["anio_ing_carr_ori"]
g.head()

if "dif_ing_ult_mat" not in df.columns:
    print(df.shape)
    df = pd.merge(
        df,
        g[["mrun", "codigo_unico", "dif_ing_ult_mat"]],
        how="left"
    )
    print(df.shape)

df["dif_ing_ult_mat"].value_counts("%") * 100

## Edad

# Se pisa año de ingreso por el minimo
print(df.shape)
df = pd.merge(
    df.drop(columns=["anio_ing_carr_ori"]),
    g[["mrun", "codigo_unico", "anio_ing_carr_ori"]],
    how="left"
)
df.shape

df["fecha_n"] = pd.to_datetime([
    date(int(str(d)[:4]), int(str(d)[-2:]), 15)
    for d in df["fec_nac_alu"].astype(int)
])

df["Edad"] = [
    date(y, 1, 1).year - fn.year - (
        (date(y, 1, 1).month, date(y, 1, 1).day) < (fn.month, fn.day)
    )
    for fn, y in zip(df["fecha_n"], df["anio_ing_carr_ori"])
]
df.shape

## Valor arancel

valores_uf_dic_anio_anterior = {
    2014: 24627.10,
    2015: 25629.09,
    2016: 26347.98,
    2017: 26798.14,
    2018: 27565.79,
}

df["valor_arancel"] = [
    int(
        valores_uf_dic_anio_anterior[a] * va
    ) if fv == "Monto en UF" else va
    for a, fv, va in zip(
        df["cat_periodo"], df["formato_valores"], df["valor_arancel"]
    )
]

g = df.groupby(["mrun", "codigo_unico"]).agg({
    "valor_arancel": "mean",
}).reset_index()

print(df.shape)
df = pd.merge(
    df.drop(columns=["valor_arancel"]),
    g[["mrun", "codigo_unico", "valor_arancel"]],
    how="left"
)
print(df.shape)

## Decil
- Se usa el valor de decil más reciente

# Se comprueba que hay gente con más de un decil
df.groupby(["mrun"]).agg({"decil":pd.Series.nunique}).reset_index().sort_values(by=["decil"], ascending=False).head()

def fill_decil(data, mrun):
    tmp = data[data["mrun"]==mrun][["cat_periodo", "decil"]].dropna()
    decil = np.nan
    
    if tmp.empty == False:
        decil = tmp.sort_values(
            by=["cat_periodo"], ascending=False
        )["decil"].iloc[0]
        
    return decil

print(df["decil"].isna().sum())
df["decil"] = [
    fill_decil(df, r)
    for r in tqdm(df["mrun"], total=df.shape[0])
]
print(df["decil"].isna().sum())

df.to_csv(f"{BASE}tmp_data/decil_imput.csv.gz", compression="gzip", index=False)

df = pd.read_csv(f"{BASE}tmp_data/decil_imput.csv.gz")

# Se comprueba que solo hay 1 valor de decil por persona
df.groupby(["mrun"]).agg({"decil":pd.Series.nunique}).reset_index().sort_values(by=["decil"], ascending=False).head()

df["decil"] = df["decil"].astype("category")

# Registros únicos

- Se crea 1 registro por combinación `mrun` - `codigo_unico`.

for c in ["cat_periodo", "anio_ing_carr_ori", "fecha_n", "fec_nac_alu"]:
    if c in df.columns:
        df = df.drop(columns=[c])
        
df.shape

# Se mantiene el registro donde esté como desertor
t = df.drop_duplicates().sort_values(by=["mrun", "codigo_unico", "desertor_1"], ascending=[False, False, False]).drop_duplicates(subset=["mrun", "codigo_unico"], keep="first").copy()
t = t.dropna()
t.shape

# No debe haber mrun-codigo_unico repetidos
assert t[["mrun", "codigo_unico"]].value_counts().reset_index()[0].max() == 1

# Se revisa dif
t["dif_ing_ult_mat"].value_counts("%") * 100

print("Desertores")
print(t[t["desertor_1"]==1]["dif_ing_ult_mat"].value_counts("%") * 100)
print("\nNo desetores")
print(t[t["desertor_1"]==0]["dif_ing_ult_mat"].value_counts("%") * 100)

t["desertor_1"].value_counts("%")

# Se evalúa decil
t["decil"].value_counts("%") * 100

# Pre procesamiento resto de variables
## Tipo enseñanza

omit_dep = [
    'LOS ANDES', 'SANTIAGO ORIENTE', 'VALPARAISO', 'SANTIAGO PONIENTE',
    'COLCHAGUA', 'VALPARAÍSO', 'CAUTÍN NORTE', 'ANTOFAGASTA - TOCOPILLA'
]

t = t[~t["dependencia_colegio"].isin(omit_dep)]
t.shape

t["dependencia_colegio"] = t["dependencia_colegio"].astype(int)
t["dependencia_colegio"].unique()

t["dependencia_colegio"] = t["dependencia_colegio"].map({
    1: 'Municipal',
    2: 'Municipal',
    3: 'Particular Subvencionado',
    4: 'Particular Pagado',
    5: 'Otro',
    6: 'Otro'
})
t["dependencia_colegio"].value_counts()

## Dependencia Colegio

t["tipo_ensenanza_colegio"] = t["tipo_ensenanza_colegio"].map({
    0: "no_aplica",
    10: "Parvularia",
    110: "Basica",
    160: "Basica",
    161: "Basica",
    163: "Carceles",
    165: "Basica",
    167: "Basica",
    211: "Especial",
    212: "Especial",
    213: "Especial",
    214: "Especial",
    215: "Especial",
    216: "Especial",
    217: "Especial",
    299: "PIE",
    310: "HC",
    360: "HC",
    361: "HC",
    362: "Carceles",
    363: "HC",
    410: "TP Comercial",
    460: "TP Comercial",
    461: "TP Comercial",
    463: "TP Comercial",
    510: "TP Industrial",
    560: "TP Industrial",
    561: "TP Industrial",
    563: "TP Industrial",
    610: "TP Tecnica",
    660: "TP Tecnica",
    661: "TP Tecnica",
    663: "TP Tecnica",
    710: "Otro",
    760: "Otro",
    761: "Otro",
    763: "Otro",
    810: "Otro",
    860: "Otro",
    863: "Otro",
    910: "Otro",
    963: "Otro"                     
})

t["tipo_ensenanza_colegio"].value_counts()

## Región colegio

t["region_colegio"] = t["region_colegio"].astype(int)
t["region_colegio"].unique()

t["region_colegio"] = t["region_colegio"].map({
    1: "Tarapacá",
    2: "Antofagasta",
    3: "Atacama",
    4: "Coquimbo",
    5: "Valparaíso",
    6: "Lib. Gral B. O'Higgins",
    7: "Maule",
    8: "Biobío",
    9: "La Araucanía",
    10: "Los Lagos",
    11: "Aysén",
    12: "Magallanes",
    13: "Metropolitana",
    14: "Los Ríos",
    15: "Arica y Parinacota",
    16: "Ñuble"
})
t["region_colegio"].value_counts()

t["misma_region"] = np.where(t["region_colegio"]==t["region_sede"], 1, 0)
t["misma_region"].value_counts("%") * 100

## Genero, desertor, beneficios

t["gen_alu"] = t["gen_alu"].replace([1, 2], ["hombre", "mujer"])
t["desertor_1"] = t["desertor_1"].astype(int)
t["fscu"] = t["fscu"].astype(int)
t["gratuidad"] = t["gratuidad"].astype(int)
t["beca"] = t["beca"].astype(int)
t["gen_alu"].value_counts()

## Duraciones

t["dur_estudio_carr"] = t["dur_estudio_carr"].astype(int)
t["dur_proceso_tit"] = t["dur_proceso_tit"].astype(int)

## Instituciones

def iter_df_read(file, col_name, condition_list, chunksize=1000, **kwargs):
    iter_csv = pd.read_csv(file, iterator=True, chunksize=chunksize, **kwargs)
    return pd.concat([chunk[chunk[col_name].isin(condition_list)] for chunk in iter_csv])

cols = ['tipo_inst_1', 'codigo_unico', 'nomb_inst']
tipo_inst = ['Centros de Formación Técnica', 'Institutos Profesionales']

instituciones = pd.concat([
    iter_df_read(f, "tipo_inst_1", tipo_inst, sep=";", usecols=cols)
    for f in tqdm(glob(f"{BASE}/raw_data/post_2015/20220719_Matrícula_Ed_Superior_*.csv"), total=8)
]).drop_duplicates().reset_index(drop=True)

assert instituciones.groupby(["codigo_unico"]).agg({"nomb_inst": pd.Series.nunique}).reset_index()["nomb_inst"].max() == 1

if "nomb_inst" not in t.columns:     
    t_inicial = t.shape[0]
    t = pd.merge(
        t,
        instituciones,
        how="left"
    )

    assert t.shape[0] == t_inicial
    
    cft_estatal = [
        'CFT DE LA REGION DE ANTOFAGASTA',
        'CFT DE LA REGION DE ARICA Y PARINACOTA',
        'CFT DE LA REGION DE ATACAMA',
        'CFT DE LA REGION DE AYSEN DEL GENERAL CARLOS IBAÑEZ DEL CAMPO',
        'CFT DE LA REGION DE COQUIMBO',
        'CFT DE LA REGION DE LA ARAUCANIA',
        'CFT DE LA REGION DE LOS LAGOS',
        'CFT DE LA REGION DE LOS RIOS',
        'CFT DE LA REGION DE MAGALLANES Y ANTARTICA CHILENA',
        'CFT DE LA REGION DE TARAPACA',
        'CFT DE LA REGION DE VALPARAISO',
        'CFT DE LA REGION DEL BIO BIO',
        "CFT DE LA REGION DEL LIBERTADOR GENERAL BERNARDO O'HIGGINS",
        'CFT DE LA REGION DEL MAULE',
        'CFT DE LA REGION METROPOLITANA DE SANTIAGO'
    ]
    
    ip_nac = [
        'IP AIEP',
        'IP DE CHILE',
        'IP DUOC UC',
        'IP INACAP',
        'IP IPG',
        'IP LATINOAMERICANO DE COMERCIO EXTERIOR - IPLACEX',
        'IP SANTO TOMAS',
    ]

    t["sub_tipo_inst"] = np.where(
        t["nomb_inst"].isin(cft_estatal),
        "cft_estatal",
        np.where(
            t["tipo_inst_1"] == "Centros de Formación Técnica",
            "cft_privado",
            np.where(
                t["nomb_inst"].isin(ip_nac),
                "ip_privado_precencia_nacional",
                "ip_privado"
            )
        )
    )

    assert t.shape[0] == t_inicial
    
t["sub_tipo_inst"].value_counts(dropna=False)

# Sacar cols auxiliares

t_aux = t.copy().drop(columns=[
    "region_colegio", "nomb_inst", "formato_valores"
]).reset_index(drop=True)

dummy_aux = pd.get_dummies(
    t_aux.drop(columns=["codigo_unico", "mrun"])
).reset_index(drop=True)

dummy_aux = pd.concat([
    t_aux[["mrun", "codigo_unico"]],
    dummy_aux
], axis=1)

dummy_aux.columns = [
    unidecode(c.replace(" ", "_").replace(".", "").replace("'", "").lower())
    for c in dummy_aux.columns
]
dummy_aux.to_csv(f"{BASE}work_data/matriz_desercion_5_dummy_con_id.csv.gz", index=False, compression="gzip")

t = t.drop(columns=["codigo_unico", "mrun", "region_colegio", "nomb_inst", "formato_valores"])

# No dummy

t.head()

t.to_csv(f"{BASE}work_data/matriz_desercion_5_no_dummy.csv.gz", index=False, compression="gzip")

# Dummys

t.info()

dummy = pd.get_dummies(t, drop_first=True)
dummy.head()

dummy.columns = [
    unidecode(c.replace(" ", "_").replace(".", "").replace("'", "").lower())
    for c in dummy.columns
]

dummy.columns

- En género elimina hombre.
- En tipo institución elimina CFT.
- En `region_sede` elimina Antofagasta.
- En `area_conocimiento` elimina Administración y Comercio.
- En ambas acreditaciones elimina Acreditada.
- En `tipo_ensenanza_colegio` elimina HC.
- En `dependencia_colegio` elimina Municipal.

dummy.to_csv(f"{BASE}work_data/matriz_desercion_5_dummy.csv.gz", index=False, compression="gzip")

# Dummy sin drop first

dummy_ohe = pd.get_dummies(t)
dummy_ohe.head()

dummy_ohe.columns = [
    unidecode(c.replace(" ", "_").replace(".", "").replace("'", "").lower())
    for c in dummy_ohe.columns
]

dummy_ohe.columns

dummy_ohe.to_csv(f"{BASE}work_data/matriz_desercion_5_dummy_ohe.csv.gz", index=False, compression="gzip")
