==> notebooks/9-seleccion-atributos.ipynb <==
import pandas as pd
import numpy as np
from tqdm.notebook import tqdm
from glob import glob
from utils import iter_df_read

BASE = "/media/giani/Gianicosas/Magister/Proyecto/"

df = pd.read_csv(f"{BASE}work_data/matriz_desercion_5_dummy_con_id.csv.gz")
df.head()

shap_best = [
    "dif_ing_ult_mat", "valor_arancel", "beca", "prom_notas_media", "gen_alu_hombre",
    "dur_proceso_tit", "edad",  "jornada_vespertino",
    "acreditada_carr_acreditada", "dur_estudio_carr",
    "dependencia_colegio_municipal", "gratuidad", "region_sede_metropolitana", 
    "decil_40", "sub_tipo_inst_ip_privado_precencia_nacional", "decil_90",
    "area_conocimiento_arte_y_arquitectura", "area_conocimiento_tecnologia",
    "jornada_diurno", 'acreditada_inst_acreditada'
]

tipo_colegio = [
    'tipo_ensenanza_colegio_hc', 'tipo_ensenanza_colegio_tp_comercial',
    'tipo_ensenanza_colegio_tp_industrial', 'tipo_ensenanza_colegio_tp_tecnica'  
]

# Matriculas 2015

cols_mat = [
    'cat_periodo', 'mrun', "codigo_unico", "anio_ing_carr_ori", "tipo_inst_1"
]

instituciones = ['Institutos Profesionales', 'Centros de Formación Técnica']

matriculados = pd.concat([
    iter_df_read(f, "tipo_inst_1", instituciones, sep=";", usecols=cols_mat)
    for f in tqdm(
        glob(f"{BASE}/raw_data/post_2015/20220719_Matrícula_Ed_Superior_*.csv"),
        total=8
    )
]).dropna(subset="mrun").drop_duplicates().reset_index(drop=True)

matriculados = matriculados[
    (matriculados["anio_ing_carr_ori"] == 2015)
]

print(f"Cantidad de registros: {matriculados.shape[0]}\n")
print(f"Cantidad de nulos: \n{matriculados.isna().sum()}")
matriculados.head()

df_chico = df[shap_best + tipo_colegio + ["desertor_1", "mrun", "codigo_unico"]]

df_left = pd.merge(
    df_chico,
    matriculados[["codigo_unico", "mrun"]].drop_duplicates(),
    how="left"
)

df_chico.shape, matriculados.shape, df_left.shape

df_left["desertor_1"].value_counts()

df_left.drop(columns=["mrun", "codigo_unico"]).to_csv(
    f"{BASE}work_data/matriz_desercion_5_mats_2015_dep.csv.gz",
    index=False,
    compression="gzip"
)
